# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Deployer API
  version: v1
  description: >
    Machine-readable reference for the Deployer admin API and runner agent API.
    Admin routes are intended to be protected by an upstream authorization
    gateway; Deployer does not implement local admin password login. Browser
    write requests also send X-Deployer-CSRF: "1". Agent routes use bearer
    tokens created through the runner API. Private control-plane routes use
    scoped Deployer service tokens and are not customer-facing endpoints.
servers:
  - url: https://deployer.example.com
    description: Example production URL
  - url: http://127.0.0.1:9090
    description: Default local URL
...
paths:
...
  /api/internal/v1/capabilities:
    get:
      tags: [internal]
      summary: Negotiate hosting manifest, runner, runtime, resource, and failure-code capabilities
      operationId: getInternalCapabilities
      security: [{serviceBearerAuth: []}]
      responses:
        "200":
          description: Hosting execution capabilities.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HostingCapabilities"}
  /api/internal/v1/runners:
    get:
      tags: [internal]
      summary: List hosting-only runner capabilities and free capacity
      operationId: listInternalHostingRunners
      security: [{serviceBearerAuth: []}]
      responses:
        "200":
          description: Hosting runners; legacy trusted-admin runners are excluded.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/HostingRunnerState"}
...
  /api/internal/v1/settings/kill-switch:
    put:
      tags: [internal]
      summary: Change the global hosting execution kill switch
      description: Requires hosting:admin. Enabling cancels queued work and durably requests cancellation of leased work.
      operationId: setInternalGlobalKillSwitch
      security: [{serviceBearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/KillSwitchInput"}
      responses:
        "200":
          description: Audited global kill-switch state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/KillSwitchState"}
        "400": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    agentBearerAuth:
      type: http
      scheme: bearer
    serviceBearerAuth:
      type: http
      scheme: bearer
      description: Hashed scoped service token. Supported scopes are deployments:read, deployments:write, projects:write, and hosting:admin.
    hostingAgentBearerAuth:
      type: http
      scheme: bearer
      description: Dedicated hashed hosting-runner credential; never interchangeable with a legacy runner or service token.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Scoped to issuer and operation. Reuse with a different canonical request hash returns idempotency_conflict.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{8,128}$'
...
  responses:
    Error:
      description: JSON error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          description: Stable machine-readable code; clients never parse error text.
          enum:
            - authentication_required
            - invalid_agent_token
            - service_token_required
            - invalid_service_token
            - service_token_not_found
            - insufficient_scope
            - invalid_idempotency_key
            - idempotency_conflict
            - idempotency_in_progress
            - invalid_deployment
            - external_deployment_conflict
            - deployment_not_found
            - manifest_digest_mismatch
            - artifact_digest_mismatch
            - source_fetch_failed
            - project_suspended
            - execution_disabled
            - runner_capacity_unavailable
            - release_not_found
            - release_not_healthy
            - proxy_unavailable
            - proxy_rejected
            - unsupported_media_type
            - manifest_signature_required
            - invalid_manifest_signature
            - manifest_signature_expired
            - invalid_manifest
            - provision_failed
            - csrf_required
            - forbidden
            - validation_error
            - method_not_allowed
            - not_found
            - conflict
            - project_busy
            - build_not_running
            - build_cancel_failed
            - payload_too_large
            - internal_error
            - project_not_found
            - build_not_found
            - runner_not_found
            - job_not_found
            - job_forbidden
            - artifact_not_found
            - artifact_unavailable
            - artifact_unmanaged
            - artifact_too_large
            - snapshot_unavailable
            - snapshot_unmanaged
            - invalid_agent_completion
...
    HostingWorkloadLimits:
      type: object
      additionalProperties: false
      required: [cpu_millis, ram_bytes, disk_bytes, pids]
      properties:
        cpu_millis: {type: integer, format: int64, minimum: 1}
        ram_bytes: {type: integer, format: int64, minimum: 1}
        disk_bytes: {type: integer, format: int64, minimum: 1}
        pids: {type: integer, format: int64, minimum: 1}
...
    KillSwitchInput:
      type: object
      additionalProperties: false
      required: [enabled, reason]
      properties:
        enabled: {type: boolean}
        reason: {type: string, minLength: 3, maxLength: 500, pattern: '^[^\r\n]+$'}
    KillSwitchState:
      type: object
      required: [scope, enabled, reason, updated_at]
      properties:
        scope: {type: string, enum: [global, project]}
        external_project_id: {type: string}
        enabled: {type: boolean}
        reason: {type: string, description: Redacted audit reason.}
        updated_at: {type: string, format: date-time}
    Capacity:
      type: object
      required: [cpu_millis, ram_bytes, disk_bytes, pids]
      properties:
        cpu_millis: {type: integer, format: int64, minimum: 0}
        ram_bytes: {type: integer, format: int64, minimum: 0}
        disk_bytes: {type: integer, format: int64, minimum: 0}
        pids: {type: integer, format: int64, minimum: 0}
    HostingRunnerState:
      type: object
      required: [id, name, labels, protocol_version, manifest_versions, runtime_versions, capacity, free, reserve, draining, status]
      properties:
        id: {type: integer, format: int64}
        name: {type: string}
        labels: {type: array, items: {type: string}}
        protocol_version: {type: string, const: v1}
        manifest_versions: {type: array, items: {type: string}}
        runtime_versions: {type: array, items: {type: string, enum: ['20', '22']}}
        capacity: {$ref: "#/components/schemas/Capacity"}
        free: {$ref: "#/components/schemas/Capacity"}
        reserve: {$ref: "#/components/schemas/Capacity"}
        draining: {type: boolean}
        status: {type: string, enum: [online, offline]}
        last_seen: {type: string, format: date-time}
...
    HostingCapabilities:
      type: object
      required: [api_version, manifest_versions, runner_protocol_versions, node_versions, runtime_kinds, resource_profiles, failure_codes]
      properties:
        api_version: {type: string, const: v1}
        manifest_versions: {type: array, items: {type: string, const: v1}}
        runner_protocol_versions: {type: array, items: {type: string, const: v1}}
        node_versions: {type: array, items: {type: string, enum: ['20', '22']}}
        runtime_kinds: {type: array, items: {type: string, enum: [static, node]}}
        resource_profiles:
          type: object
          additionalProperties: {$ref: "#/components/schemas/HostingWorkloadLimits"}
        failure_codes: {type: array, items: {type: string}}
...